quant_scalp_bot/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
pandas==2.2.2
python-telegram-bot[job-queue]==21.3
